+++ firestore-stripe-invoices/functions/src/logs.ts
@@ -14,34 +14,33 @@
  * limitations under the License.
  */
 
+import { logger } from 'firebase-functions';
 import Stripe from 'stripe';
 import { InvoicePayload } from './interfaces';
 
 export function startInvoiceCreate() {
-  console.log('🙂 Received new invoice, starting processing');
+  logger.log('🙂 Received new invoice, starting processing');
 }
 
 export function startInvoiceUpdate(eventType: string) {
-  console.log(
-    `🙂 Received new invoice event ${eventType}, starting processing`
-  );
+  logger.log(`🙂 Received new invoice event ${eventType}, starting processing`);
 }
 
 export function incorrectPayload(payload: InvoicePayload) {
   if (!payload.items.length) {
-    console.error(
+    logger.error(
       new Error('😞[Error] Missing at least one line item in items[]')
     );
   }
   if (!payload.email && !payload.uid) {
-    console.error(
+    logger.error(
       new Error(
         '😞[Error] Missing either a customer email address or Firebase Authentication uid'
       )
     );
   }
   if (payload.email && payload.uid) {
-    console.error(
+    logger.error(
       new Error(
         '😞[Error] Only either email or uid is permitted, you specified both.'
       )
@@ -50,27 +49,27 @@ export function incorrectPayload(payload: InvoicePayload) {
 }
 
 export function noEmailForUser(uid: string) {
-  console.error(
+  logger.error(
     new Error(`😞[Error] User [${uid}] is missing an email address.`)
   );
 }
 
 export function stripeError(err: Stripe.StripeCardError) {
-  console.error(
+  logger.error(
     new Error('😞[Error] Error when making a request to the Stripe API:'),
     err
   );
 }
 
 export function invoiceCreatedError(invoice?: Stripe.Invoice) {
-  console.error(
+  logger.error(
     new Error('😞[Error] Error when creating the invoice:'),
     invoice
   );
 }
 
 export function customerCreated(id: string, livemode: boolean) {
-  console.log(
+  logger.log(
     `👤 Created a new customer: https://dashboard.stripe.com${
       livemode ? '' : '/test'
     }/customers/${id}`
@@ -78,7 +77,7 @@ export function customerCreated(id: string, livemode: boolean) {
 }
 
 export function customerRetrieved(id: string, livemode: boolean) {
-  console.log(
+  logger.log(
     `🙋 Found existing customer by email: https://dashboard.stripe.com${
       livemode ? '' : '/test'
     }/customers/${id}`
@@ -86,7 +85,7 @@ export function customerRetrieved(id: string, livemode: boolean) {
 }
 
 export function invoiceCreated(id: string, livemode: boolean) {
-  console.log(
+  logger.log(
     `🧾 Created invoice: https://dashboard.stripe.com${
       livemode ? '' : '/test'
     }/invoices/${id}`
@@ -98,11 +97,11 @@ export function invoiceSent(
   email: string,
   hostedInvoiceUrl: string
 ) {
-  console.log(`📧 Sent invoice ${id} to ${email}: ${hostedInvoiceUrl}`);
+  logger.log(`📧 Sent invoice ${id} to ${email}: ${hostedInvoiceUrl}`);
 }
 
 export function badSignature(err: Error) {
-  console.error(
+  logger.error(
     '😞[Error] Webhook signature verification failed. Is your Stripe webhook secret parameter configured correctly?',
     err
   );
@@ -117,11 +116,11 @@ export function malformedEvent(event: Stripe.Event) {
     err = new Error('Could not find event.type');
   }
 
-  console.error('😞[Error] Malformed event', err);
+  logger.error('😞[Error] Malformed event', err);
 }
 
 export function ignoreEvent(eventType: string) {
-  console.log(
+  logger.log(
     `🙈 Ignoring event "${eventType}" because it because it isn't a relevant part of the invoice lifecycle`
   );
 }
@@ -130,7 +129,7 @@ export function unexpectedInvoiceAmount(
   numInvoices: number,
   invoiceId: string
 ) {
-  console.error(
+  logger.error(
     '😞[Error] could not find invoice',
     new Error(
       `Expected 1 invoice with ID "${invoiceId}", but found ${numInvoices}`
@@ -143,7 +142,7 @@ export function statusUpdateComplete(
   newStatus: string,
   eventType: string
 ) {
-  console.log(
+  logger.log(
     `🙂 Updated invoice "${invoiceId}" to status "${newStatus}" on event type "${eventType}"`
   );
 }